+++ services/auth/go/controller/change_user_password_test.go
@@ -75,6 +75,16 @@ func TestChangeUserPassword(t *testing.T) { //nolint:maintidx
 					),
 				).Return(userID, nil)
 
+				mock.EXPECT().DeleteRefreshTokens(
+					gomock.Any(),
+					userID,
+				).Return(nil)
+
+				mock.EXPECT().DeleteOAuth2RefreshTokensByUserID(
+					gomock.Any(),
+					userID,
+				).Return(nil)
+
 				return mock
 			},
 			jwtTokenFn: nil,
@@ -246,6 +256,16 @@ func TestChangeUserPassword(t *testing.T) { //nolint:maintidx
 					),
 				).Return(userID, nil)
 
+				mock.EXPECT().DeleteRefreshTokens(
+					gomock.Any(),
+					userID,
+				).Return(nil)
+
+				mock.EXPECT().DeleteOAuth2RefreshTokensByUserID(
+					gomock.Any(),
+					userID,
+				).Return(nil)
+
 				return mock
 			},
 			jwtTokenFn: jwtTokenFn,
@@ -408,6 +428,220 @@ func TestChangeUserPassword(t *testing.T) { //nolint:maintidx
 				Error:   "password-too-short",
 				Message: "Password is too short",
 				Status:  400,
+			},
+			expectedJWT:       nil,
+			getControllerOpts: []getControllerOptsFunc{},
+		},
+
+		{
+			name:   "ticket - error deleting refresh tokens",
+			config: getConfig,
+			db: func(ctrl *gomock.Controller) controller.DBClient {
+				mock := mock.NewMockDBClient(ctrl)
+
+				mock.EXPECT().GetUserByTicket(
+					gomock.Any(),
+					sql.Text("passwordReset:ticket"),
+				).Return(sql.AuthUser{ //nolint:exhaustruct
+					ID:    userID,
+					Email: sql.Text("user@acme.local"),
+				}, nil)
+
+				mock.EXPECT().UpdateUserChangePassword(
+					gomock.Any(),
+					testhelpers.GomockCmpOpts(
+						sql.UpdateUserChangePasswordParams{
+							ID:           userID,
+							PasswordHash: sql.Text("password"),
+						},
+						cmpopts.IgnoreFields(
+							sql.UpdateUserChangePasswordParams{}, //nolint:exhaustruct
+							"PasswordHash",
+						),
+					),
+				).Return(userID, nil)
+
+				mock.EXPECT().DeleteRefreshTokens(
+					gomock.Any(),
+					userID,
+				).Return(pgx.ErrTxClosed)
+
+				return mock
+			},
+			jwtTokenFn: nil,
+			request: api.ChangeUserPasswordRequestObject{
+				Body: &api.ChangeUserPasswordJSONRequestBody{
+					NewPassword: "password",
+					Ticket:      new("passwordReset:ticket"),
+				},
+			},
+			expectedResponse: controller.ErrorResponse{
+				Error:   "internal-server-error",
+				Message: "Internal server error",
+				Status:  500,
+			},
+			expectedJWT:       nil,
+			getControllerOpts: []getControllerOptsFunc{},
+		},
+
+		{
+			name:   "ticket - error deleting oauth2 refresh tokens",
+			config: getConfig,
+			db: func(ctrl *gomock.Controller) controller.DBClient {
+				mock := mock.NewMockDBClient(ctrl)
+
+				mock.EXPECT().GetUserByTicket(
+					gomock.Any(),
+					sql.Text("passwordReset:ticket"),
+				).Return(sql.AuthUser{ //nolint:exhaustruct
+					ID:    userID,
+					Email: sql.Text("user@acme.local"),
+				}, nil)
+
+				mock.EXPECT().UpdateUserChangePassword(
+					gomock.Any(),
+					testhelpers.GomockCmpOpts(
+						sql.UpdateUserChangePasswordParams{
+							ID:           userID,
+							PasswordHash: sql.Text("password"),
+						},
+						cmpopts.IgnoreFields(
+							sql.UpdateUserChangePasswordParams{}, //nolint:exhaustruct
+							"PasswordHash",
+						),
+					),
+				).Return(userID, nil)
+
+				mock.EXPECT().DeleteRefreshTokens(
+					gomock.Any(),
+					userID,
+				).Return(nil)
+
+				mock.EXPECT().DeleteOAuth2RefreshTokensByUserID(
+					gomock.Any(),
+					userID,
+				).Return(pgx.ErrTxClosed)
+
+				return mock
+			},
+			jwtTokenFn: nil,
+			request: api.ChangeUserPasswordRequestObject{
+				Body: &api.ChangeUserPasswordJSONRequestBody{
+					NewPassword: "password",
+					Ticket:      new("passwordReset:ticket"),
+				},
+			},
+			expectedResponse: controller.ErrorResponse{
+				Error:   "internal-server-error",
+				Message: "Internal server error",
+				Status:  500,
+			},
+			expectedJWT:       nil,
+			getControllerOpts: []getControllerOptsFunc{},
+		},
+
+		{
+			name:   "auth header - error deleting refresh tokens",
+			config: getConfig,
+			db: func(ctrl *gomock.Controller) controller.DBClient {
+				mock := mock.NewMockDBClient(ctrl)
+
+				mock.EXPECT().GetUser(
+					gomock.Any(),
+					userID,
+				).Return(sql.AuthUser{ //nolint:exhaustruct
+					ID:    userID,
+					Email: sql.Text("user@acme.local"),
+				}, nil)
+
+				mock.EXPECT().UpdateUserChangePassword(
+					gomock.Any(),
+					testhelpers.GomockCmpOpts(
+						sql.UpdateUserChangePasswordParams{
+							ID:           userID,
+							PasswordHash: sql.Text("password"),
+						},
+						cmpopts.IgnoreFields(
+							sql.UpdateUserChangePasswordParams{}, //nolint:exhaustruct
+							"PasswordHash",
+						),
+					),
+				).Return(userID, nil)
+
+				mock.EXPECT().DeleteRefreshTokens(
+					gomock.Any(),
+					userID,
+				).Return(pgx.ErrTxClosed)
+
+				return mock
+			},
+			jwtTokenFn: jwtTokenFn,
+			request: api.ChangeUserPasswordRequestObject{
+				Body: &api.ChangeUserPasswordJSONRequestBody{
+					NewPassword: "password",
+					Ticket:      nil,
+				},
+			},
+			expectedResponse: controller.ErrorResponse{
+				Error:   "internal-server-error",
+				Message: "Internal server error",
+				Status:  500,
+			},
+			expectedJWT:       nil,
+			getControllerOpts: []getControllerOptsFunc{},
+		},
+
+		{
+			name:   "auth header - error deleting oauth2 refresh tokens",
+			config: getConfig,
+			db: func(ctrl *gomock.Controller) controller.DBClient {
+				mock := mock.NewMockDBClient(ctrl)
+
+				mock.EXPECT().GetUser(
+					gomock.Any(),
+					userID,
+				).Return(sql.AuthUser{ //nolint:exhaustruct
+					ID:    userID,
+					Email: sql.Text("user@acme.local"),
+				}, nil)
+
+				mock.EXPECT().UpdateUserChangePassword(
+					gomock.Any(),
+					testhelpers.GomockCmpOpts(
+						sql.UpdateUserChangePasswordParams{
+							ID:           userID,
+							PasswordHash: sql.Text("password"),
+						},
+						cmpopts.IgnoreFields(
+							sql.UpdateUserChangePasswordParams{}, //nolint:exhaustruct
+							"PasswordHash",
+						),
+					),
+				).Return(userID, nil)
+
+				mock.EXPECT().DeleteRefreshTokens(
+					gomock.Any(),
+					userID,
+				).Return(nil)
+
+				mock.EXPECT().DeleteOAuth2RefreshTokensByUserID(
+					gomock.Any(),
+					userID,
+				).Return(pgx.ErrTxClosed)
+
+				return mock
+			},
+			jwtTokenFn: jwtTokenFn,
+			request: api.ChangeUserPasswordRequestObject{
+				Body: &api.ChangeUserPasswordJSONRequestBody{
+					NewPassword: "password",
+					Ticket:      nil,
+				},
+			},
+			expectedResponse: controller.ErrorResponse{
+				Error:   "internal-server-error",
+				Message: "Internal server error",
+				Status:  500,
 			},
 			expectedJWT:       nil,
 			getControllerOpts: []getControllerOptsFunc{},